// datum-erp-workspace/apps/mfe/financeApp/src/app/remote-entry/model/paccountlistcomponent.model.ts
export interface ACCOUNTLIST {
    id: number;
    description: string;
  }
  
  export interface ACCOUNTPOPUP{
    alias: string;
    name: string;
    id: number;
  }
  
  export interface ACCOUNTBYLIST{
    // listID: number;
    // accountID: number;
    // branchID: number;
    alias: string;
    name: string;
    id: number;
  }
  
  